Don’t use Fuel for downloading files, it does not follow redirects
Of course that is half incorrect but when trying to download something
to a File, Fuel doesn’t handle that correctly on a redirect: the second
request always has all the data in memory, which for multi-100-MB files
is probably not what we want.

The solution is to use HttpUrlConnection and friends directly.